+++ src/main/java/org/opensearch/knn/index/mapper/KNNVectorFieldMapper.java
@@ -58,6 +58,7 @@ import static org.opensearch.knn.common.KNNValidationUtil.validateVectorDimensio
 import static org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil.createKNNMethodContextFromLegacy;
 import static org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil.createStoredFieldForByteVector;
 import static org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil.createStoredFieldForFloatVector;
+import static org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil.useFullFieldNameValidation;
 import static org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil.validateIfCircuitBreakerIsNotTriggered;
 import static org.opensearch.knn.index.mapper.KNNVectorFieldMapperUtil.validateIfKNNPluginEnabled;
 import static org.opensearch.knn.index.mapper.ModelFieldMapper.UNSET_MODEL_DIMENSION_IDENTIFIER;
@@ -93,6 +94,7 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
      */
     public static class Builder extends ParametrizedFieldMapper.Builder {
         protected Boolean ignoreMalformed;
+        protected final boolean isDerivedSourceEnabled;
 
         protected final Parameter<Boolean> stored = Parameter.storeParam(m -> toType(m).stored, false);
         protected final Parameter<Boolean> hasDocValues = Parameter.docValuesParam(m -> toType(m).hasDocValues, true);
@@ -200,13 +202,15 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
             ModelDao modelDao,
             Version indexCreatedVersion,
             KNNMethodConfigContext knnMethodConfigContext,
-            OriginalMappingParameters originalParameters
+            OriginalMappingParameters originalParameters,
+            boolean isDerivedSourceEnabled
         ) {
             super(name);
             this.modelDao = modelDao;
             this.indexCreatedVersion = indexCreatedVersion;
             this.knnMethodConfigContext = knnMethodConfigContext;
             this.originalParameters = originalParameters;
+            this.isDerivedSourceEnabled = isDerivedSourceEnabled;
         }
 
         @Override
@@ -237,7 +241,9 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
 
         @Override
         public KNNVectorFieldMapper build(BuilderContext context) {
-            validateFullFieldName(context);
+            if (useFullFieldNameValidation(indexCreatedVersion)) {
+                validateFullFieldName(context);
+            }
 
             final MultiFields multiFieldsBuilder = this.multiFieldsBuilder.build(this, context);
             final CopyTo copyToBuilder = copyTo.build();
@@ -258,7 +264,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                     modelDao,
                     indexCreatedVersion,
                     originalParameters,
-                    knnMethodConfigContext
+                    knnMethodConfigContext,
+                    isDerivedSourceEnabled
                 );
             }
 
@@ -280,7 +287,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                     ignoreMalformed,
                     stored.get(),
                     hasDocValues.get(),
-                    originalParameters
+                    originalParameters,
+                    isDerivedSourceEnabled
                 );
             }
 
@@ -301,7 +309,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                     metaValue,
                     knnMethodConfigContext,
                     createLuceneFieldMapperInput,
-                    originalParameters
+                    originalParameters,
+                    isDerivedSourceEnabled
                 );
             }
 
@@ -315,7 +324,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                 ignoreMalformed,
                 stored.getValue(),
                 hasDocValues.getValue(),
-                originalParameters
+                originalParameters,
+                isDerivedSourceEnabled
             );
         }
 
@@ -363,7 +373,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
                 modelDaoSupplier.get(),
                 parserContext.indexVersionCreated(),
                 null,
-                null
+                null,
+                KNNSettings.isKNNDerivedSourceEnabled(parserContext.getSettings())
             );
             builder.parse(name, parserContext, node);
             builder.setOriginalParameters(new OriginalMappingParameters(builder));
@@ -578,6 +589,7 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
     // values of KNN engine Algorithms hyperparameters.
     protected Version indexCreatedVersion;
     protected Explicit<Boolean> ignoreMalformed;
+    protected final boolean isDerivedSourceEnabled;
     protected boolean stored;
     protected boolean hasDocValues;
     protected VectorDataType vectorDataType;
@@ -598,7 +610,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
         boolean stored,
         boolean hasDocValues,
         Version indexCreatedVersion,
-        OriginalMappingParameters originalMappingParameters
+        OriginalMappingParameters originalMappingParameters,
+        boolean isDerivedSourceEnabled
     ) {
         super(simpleName, mappedFieldType, multiFields, copyTo);
         this.ignoreMalformed = ignoreMalformed;
@@ -608,6 +621,7 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
         updateEngineStats();
         this.indexCreatedVersion = indexCreatedVersion;
         this.originalMappingParameters = originalMappingParameters;
+        this.isDerivedSourceEnabled = isDerivedSourceEnabled;
     }
 
     public KNNVectorFieldMapper clone() {
@@ -840,7 +854,8 @@ public abstract class KNNVectorFieldMapper extends ParametrizedFieldMapper {
             modelDao,
             indexCreatedVersion,
             knnMethodConfigContext,
-            originalMappingParameters
+            originalMappingParameters,
+            isDerivedSourceEnabled
         ).init(this);
     }