+++ src/test/java/org/opensearch/knn/index/mapper/KNNVectorFieldMapperTests.java
@@ -123,7 +123,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
             modelDao,
             CURRENT,
             null,
-            new OriginalMappingParameters(VectorDataType.DEFAULT, TEST_DIMENSION, null, null, null, null, SpaceType.UNDEFINED.getValue())
+            new OriginalMappingParameters(VectorDataType.DEFAULT, TEST_DIMENSION, null, null, null, null, SpaceType.UNDEFINED.getValue()),
+            false
         );
 
         assertEquals(10, builder.getParameters().size());
@@ -357,7 +358,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
     public void testBuilder_build_fromModel() {
         // Check that modelContext takes precedent over legacy
         ModelDao modelDao = mock(ModelDao.class);
-        KNNVectorFieldMapper.Builder builder = new KNNVectorFieldMapper.Builder("test-field-name-1", modelDao, CURRENT, null, null);
+        KNNVectorFieldMapper.Builder builder = new KNNVectorFieldMapper.Builder("test-field-name-1", modelDao, CURRENT, null, null, false);
 
         SpaceType spaceType = SpaceType.COSINESIMIL;
         int m = 17;
@@ -1176,6 +1177,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                 when(parseContext.parser()).thenReturn(createXContentParser(dataType));
 
                 utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useLuceneKNNVectorsFormat(Mockito.any())).thenReturn(true);
+                utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useFullFieldNameValidation(Mockito.any())).thenReturn(true);
 
                 OriginalMappingParameters originalMappingParameters = new OriginalMappingParameters(
                     dataType,
@@ -1197,7 +1199,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                     new Explicit<>(true, true),
                     false,
                     false,
-                    originalMappingParameters
+                    originalMappingParameters,
+                    false
                 );
                 methodFieldMapper.parseCreateField(parseContext, dimension, dataType);
 
@@ -1220,6 +1223,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                 );
 
                 utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useLuceneKNNVectorsFormat(Mockito.any())).thenReturn(false);
+                utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useFullFieldNameValidation(Mockito.any())).thenReturn(false);
 
                 document = new ParseContext.Document();
                 contentPath = new ContentPath();
@@ -1236,7 +1240,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                     new Explicit<>(true, true),
                     false,
                     false,
-                    originalMappingParameters
+                    originalMappingParameters,
+                    false
                 );
 
                 methodFieldMapper.parseCreateField(parseContext, dimension, dataType);
@@ -1284,6 +1289,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                 when(parseContext.parser()).thenReturn(createXContentParser(dataType));
 
                 utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useLuceneKNNVectorsFormat(Mockito.any())).thenReturn(true);
+                utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useFullFieldNameValidation(Mockito.any())).thenReturn(true);
 
                 OriginalMappingParameters originalMappingParameters = new OriginalMappingParameters(
                     VectorDataType.DEFAULT,
@@ -1308,7 +1314,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                     modelDao,
                     CURRENT,
                     originalMappingParameters,
-                    knnMethodConfigContext
+                    knnMethodConfigContext,
+                    false
                 );
 
                 modelFieldMapper.parseCreateField(parseContext);
@@ -1331,6 +1338,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                 );
 
                 utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useLuceneKNNVectorsFormat(Mockito.any())).thenReturn(false);
+                utilMockedStatic.when(() -> KNNVectorFieldMapperUtil.useFullFieldNameValidation(Mockito.any())).thenReturn(true);
 
                 document = new ParseContext.Document();
                 contentPath = new ContentPath();
@@ -1350,7 +1358,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                     modelDao,
                     CURRENT,
                     originalMappingParameters,
-                    knnMethodConfigContext
+                    knnMethodConfigContext,
+                    false
                 );
 
                 modelFieldMapper.parseCreateField(parseContext);
@@ -1396,7 +1405,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
             Collections.emptyMap(),
             knnMethodConfigContext,
             inputBuilder.build(),
-            originalMappingParameters
+            originalMappingParameters,
+            false
         );
         luceneFieldMapper.parseCreateField(parseContext, TEST_DIMENSION, VectorDataType.FLOAT);
 
@@ -1455,7 +1465,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
             Collections.emptyMap(),
             knnMethodConfigContext,
             inputBuilder.build(),
-            originalMappingParameters
+            originalMappingParameters,
+            false
         );
         luceneFieldMapper.parseCreateField(parseContext, TEST_DIMENSION, VectorDataType.FLOAT);
 
@@ -1502,7 +1513,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                     .dimension(TEST_DIMENSION)
                     .build(),
                 inputBuilder.build(),
-                originalMappingParameters
+                originalMappingParameters,
+                false
             )
         );
         doReturn(Optional.of(TEST_BYTE_VECTOR)).when(luceneFieldMapper)
@@ -1552,7 +1564,8 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
                     .dimension(TEST_DIMENSION)
                     .build(),
                 inputBuilder.build(),
-                originalMappingParameters
+                originalMappingParameters,
+                false
             )
         );
         doReturn(Optional.of(TEST_BYTE_VECTOR)).when(luceneFieldMapper)
@@ -1667,7 +1680,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
     public void testBuilder_whenBinaryWithLegacyKNNDisabled_thenValid() {
         // Check legacy is picked up if model context and method context are not set
         ModelDao modelDao = mock(ModelDao.class);
-        KNNVectorFieldMapper.Builder builder = new KNNVectorFieldMapper.Builder("test-field-name-1", modelDao, CURRENT, null, null);
+        KNNVectorFieldMapper.Builder builder = new KNNVectorFieldMapper.Builder("test-field-name-1", modelDao, CURRENT, null, null, false);
         builder.vectorDataType.setValue(VectorDataType.BINARY);
         builder.dimension.setValue(8);
 
@@ -1716,7 +1729,7 @@ public class KNNVectorFieldMapperTests extends KNNTestCase {
 
             // IllegalArgumentException should be thrown.
             Exception e = assertThrows(IllegalArgumentException.class, () -> {
-                new KNNVectorFieldMapper.Builder(invalidVectorFieldName, null, CURRENT, null, null).build(builderContext);
+                new KNNVectorFieldMapper.Builder(invalidVectorFieldName, null, CURRENT, null, null, false).build(builderContext);
             });
             assertTrue(e.getMessage(), e.getMessage().contains("Vector field name must not include"));
         }